+++ tests/integration/conftest.py
@@ -4,16 +4,16 @@
 # pylint: disable=unused-argument
 """Fixtures for charm tests."""
 
-import json
 import pathlib
 
 # nosec B404: subprocess usage is intentional and safe (predefined executables only).
 import subprocess  # nosec
-import textwrap
 import typing
 
 import jubilant
 import pytest
+
+from tests.integration.tinyproxy import deploy_tinyproxy
 
 
 @pytest.fixture(name="aproxy_charm_file", scope="session")
@@ -78,7 +78,7 @@ def deploy_charms_fixture(juju: jubilant.Juju, aproxy_charm_file: str, tinyproxy
         aproxy_charm_file: Path to the built aproxy charm file.
         tinyproxy_url: URL of the upstream tinyproxy.
     """
-    juju.deploy("ubuntu", base="ubuntu@22.04")
+    juju.deploy("ubuntu", base="ubuntu@24.04")
     juju.deploy(aproxy_charm_file)
     juju.integrate("ubuntu", "aproxy")
     juju.cli("config", "aproxy", f"proxy-address={tinyproxy_url}")
@@ -141,72 +141,6 @@ def aproxy_app_fixture(juju: jubilant.Juju, deploy_charms) -> App:
 
 
 @pytest.fixture(name="tinyproxy_url", scope="module")
-def deploy_tinyproxy(juju: jubilant.Juju) -> str:
-    """Deploy a tinyproxy service into the Juju model using any-charm.
-
-    Args:
-        juju: The Juju controller instance.
-
-    Returns:
-        Proxy URL (http://<unit-ip>:8888).
-    """
-    any_charm_py = textwrap.dedent(
-        """
-    import subprocess
-    import textwrap
-
-    import ops
-    from any_charm_base import AnyCharmBase
-
-
-    class AnyCharm(AnyCharmBase):
-        def __init__(self, *args):
-            super().__init__(*args)
-            self.framework.observe(self.on.install, self._on_install)
-
-        def _on_install(self, _):
-            self.unit.status = ops.MaintenanceStatus("downloading tinyproxy")
-            subprocess.check_call(["apt-get", "update"])
-            subprocess.check_call(["apt-get", "install", "-y", "tinyproxy"])
-
-            self.unit.status = ops.MaintenanceStatus("configuring tinyproxy")
-            with open("/etc/tinyproxy/tinyproxy.conf", "w") as f:
-                f.write(
-                    textwrap.dedent(
-                        \"\"\"
-                        User tinyproxy
-                        Group tinyproxy
-                        Port 8888
-                        Timeout 600
-                        DefaultErrorFile "/usr/share/tinyproxy/default.html"
-                        StatFile "/usr/share/tinyproxy/stats.html"
-                        LogFile "/var/log/tinyproxy/tinyproxy.log"
-                        LogLevel Info
-                        PidFile "/run/tinyproxy/tinyproxy.pid"
-                        MaxClients 100
-                        \"\"\"
-                    )
-                )
-
-            subprocess.check_call(["systemctl", "restart", "tinyproxy"])
-            self.unit.set_ports(8888)
-            self.unit.status = ops.ActiveStatus()
-    """
-    )
-
-    # Deploy any-charm with our custom inline tinyproxy charm
-    juju.deploy(
-        "any-charm",
-        "tinyproxy",
-        channel="latest/edge",
-        config={"src-overwrite": json.dumps({"any_charm.py": any_charm_py})},
-    )
-
-    # Wait until the service is up
-    juju.wait(jubilant.all_active, timeout=10 * 60)
-
-    # Grab unit IP
-    units = juju.status().get_units("tinyproxy")
-    leader = next(name for name, u in units.items() if u.leader)
-    unit_ip = units[leader].public_address
-    return f"http://{unit_ip}:8888"
+def get_tinyproxy_url(juju):
+    """Deploy Tinyproxy and return its URL."""
+    return deploy_tinyproxy(juju)
